refactor: consolidate ScrollableParallaxView into ParallaxScrollView component
- Moved ScrollableParallaxView functionality into main ParallaxScrollView component
- Removed separate wrapper file to reduce code duplication
- Updated imports in tab screens to use the consolidated component
- Updated documentation to reflect new import structure
- Maintains backward compatibility with ScrollableParallaxView export
