+++ src/app/(tabs)/index.tsx
@@ -4,7 +4,8 @@ import { Platform, StyleSheet, Pressable } from 'react-native';
 import { HelloWave } from '@/components/hello-wave';
 import { ThemedText } from '@/components/themed/themed-text';
 import { ThemedView } from '@/components/themed/themed-view';
-import { TabScreenWrapper, ScrollableParallaxView } from '@/components/layout';
+import { TabScreenWrapper } from '@/components/layout';
+import { ScrollableParallaxView } from '@/components/parallax-scroll-view';
 import { useScrollToTop } from '@/hooks/use-scroll-to-top';
 
 function ScrollToTopButton() {
