+++ src/components/parallax-scroll-view.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren, ReactElement } from 'react';
 import { StyleSheet } from 'react-native';
-import { forwardRef, useImperativeHandle } from 'react';
+import React, { forwardRef, useImperativeHandle, useRef, useEffect } from 'react';
 import Animated, {
   interpolate,
   useAnimatedRef,
@@ -11,6 +11,7 @@ import Animated, {
 import { ThemedView } from '@/components/themed/themed-view';
 import { useBottomTabOverflow } from '@/components/ui/tab-bar-background';
 import { useColorScheme } from '@/hooks/use-color-scheme';
+import { useScrollToTop } from '@/hooks/use-scroll-to-top';
 
 const HEADER_HEIGHT = 250;
 
@@ -89,6 +90,27 @@ const ParallaxScrollView = forwardRef<ParallaxScrollViewRef, Props>(
 ParallaxScrollView.displayName = 'ParallaxScrollView';
 
 export default ParallaxScrollView;
+
+type ParallaxScrollViewProps = React.ComponentProps<typeof ParallaxScrollView>;
+
+/**
+ * A ParallaxScrollView that automatically integrates with TabScreenWrapper's scroll-to-top functionality
+ */
+export function ScrollableParallaxView(props: ParallaxScrollViewProps) {
+  const scrollRef = useRef<ParallaxScrollViewRef>(null);
+  const scrollToTopContext = useScrollToTop();
+
+  // Register scroll handler with TabScreenWrapper
+  useEffect(() => {
+    if (scrollToTopContext && scrollRef.current) {
+      scrollToTopContext.registerScrollHandler((options) => {
+        scrollRef.current?.scrollTo(options || { y: 0, animated: true });
+      });
+    }
+  }, [scrollToTopContext]);
+
+  return <ParallaxScrollView ref={scrollRef} {...props} />;
+}
 
 const styles = StyleSheet.create({
   container: {
